requirements.txt
absl-py==1.4.0
accelerate==0.12.0
addict==2.4.0
aenum==3.1.12
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
antlr4-python3-runtime==4.9.3
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
basicsr==1.4.2
beautifulsoup4==4.12.0
blendmodes==2022
boltons==23.0.0
cachetools==5.3.0
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.1.0
clean-fid==0.1.29
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
deprecation==2.1.0
einops==0.4.1
entrypoints==0.4
facexlib==0.2.5
fastapi==0.94.0
ffmpy==0.3.0
filelock==3.10.7
filterpy==1.4.5
font-roboto==0.0.1
fonts==0.0.3
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.3.0
ftfy==6.1.1
future==0.18.3
gdown==4.7.1
gfpgan==1.3.8
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.17.1
google-auth-oauthlib==1.0.0
gradio==3.23.0
grpcio==1.53.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.3
huggingface-hub==0.13.3
idna==2.10
imageio==2.27.0
inflection==0.5.1
Jinja2==3.1.2
jsonmerge==1.8.0
jsonschema==4.17.3
kiwisolver==1.4.4
kornia==0.6.7
lark==1.1.2
lazy_loader==0.2
lightning-utilities==0.8.0
linkify-it-py==2.0.0
llvmlite==0.39.1
lmdb==1.4.0
lpips==0.1.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
multidict==6.0.4
mysql-connector-python==8.0.33
networkx==3.0
numba==0.56.4
numpy==1.23.3
oauthlib==3.2.2
omegaconf==2.2.3
open-clip-torch @ git+https://github.com/mlfoundations/open_clip.git@bb6e834e9c70d9c27d0dc3ecedeebeaeb1ffad6b
opencv-python==4.7.0.72
orjson==3.8.9
packaging==23.0
pandas==1.5.3
piexif==1.1.3
Pillow==9.4.0
protobuf==3.20.0
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.7
pydub==0.25.1
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytorch-lightning==1.9.4
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
realesrgan==0.3.0
regex==2023.3.23
requests==2.25.1
requests-oauthlib==1.3.1
resize-right==0.0.2
rfc3986==1.5.0
rsa==4.9
safetensors==0.3.0
scikit-image==0.19.2
scipy==1.10.1
semantic-version==2.10.0
sentencepiece==0.1.97
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
starlette==0.26.1
tb-nightly==2.13.0a20230402
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tifffile==2023.3.21
timm==0.6.7
tokenizers==0.13.2
toolz==0.12.0
torch==1.13.1+cu117
torchdiffeq==0.2.3
torchmetrics==0.11.4
torchsde==0.2.5
torchvision==0.14.1+cu117
tqdm==4.65.0
trampoline==0.1.2
transformers==4.25.1
typing_extensions==4.5.0
uc-micro-py==1.0.1
urllib3==1.26.15
uvicorn==0.21.1
wcwidth==0.2.6
websockets==11.0
Werkzeug==2.2.3
yapf==0.32.0
yarl==1.8.2
